refactor: 수식표현영역 검증로직 통합 (-> Validator)
