+++ src/main/java/calculator/io/CalculatorIO.java
@@ -1,7 +1,7 @@
 package calculator.io;
 
 import calculator.dto.InputDTO;
-import calculator.validation.IOValidation;
+import calculator.validation.Validator;
 import camp.nextstep.edu.missionutils.*;
 
 public class CalculatorIO {
@@ -9,7 +9,7 @@ public class CalculatorIO {
         System.out.println("덧셈할 문자열을 입력해 주세요.");
         String input = Console.readLine();
 
-        IOValidation.checkInput( input );
+        Validator.checkInput( input );
         return new InputDTO( input );
     }
