+++ src/main/java/calculator/logic/CalcLogic.java
@@ -1,6 +1,7 @@
 package calculator.logic;
 
 import calculator.dto.InputDTO;
+import calculator.validation.Validator;
 
 import java.util.Arrays;
 import java.util.List;
@@ -9,17 +10,13 @@ public class CalcLogic {
     public static List<Long> getNumbers( InputDTO input ) {
         String exp = input.getSumExp();
         String divider = input.getDividerExp();
-        String divider_regex = String.format( "[,:%s]", divider );
+        String dividerRegex = String.format( "[,:%s]", divider );
 
         // 수식 내 오류여부 확인
-        for ( String expChar: exp.replaceAll("\\d", "").split("") ) {
-            if ( expChar.length() == 0 ) continue;
-            if ( !expChar.matches( divider_regex ) )
-                throw new IllegalArgumentException("잘못된 입력입니다.");
-        }
+        Validator.checkSumExpValidity( exp, dividerRegex );
 
         // 숫자탐색
-        return Arrays.stream( exp.split( divider_regex ) )
+        return Arrays.stream( exp.split( dividerRegex ) )
                 .map( Long::parseLong )
                 .toList();
     }
